% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycalc.R
\name{mycalc_function}
\alias{mycalc_function}
\title{MyCalc Function}
\usage{
mycalc_function(x, y, fun)
}
\arguments{
\item{x}{this is the first input in the function. One can use a number, vector, matrix, or database}

\item{y}{this is the second input in the function. One can use a number, vector, matrix, or database}

\item{fun}{this is how one operationalizes the function allowing the user to add, subtract, multiply, and divide}
}
\description{
This function allows you to perform basic calculations.
}
\examples{
mycalc_function()
}
