% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycalc.R
\name{mycalc_function}
\alias{mycalc_function}
\title{MyCalc Function}
\usage{
mycalc_function(x, y, fun)
}
\arguments{
\item{calc}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to perform basic calculations.
}
\examples{
mycalc_function()
}
